Removed the second level of commands
