--- src/Dimmy.Sitecore.Plugin/ISitecoreInitialiseVersion.cs
@@ -1,11 +1,0 @@
-﻿using System.CommandLine;
-
-namespace Dimmy.Sitecore.Plugin
-{
-    public interface ISitecoreInitialiseVersion
-    {
-        string Name { get; }
-        string Description { get; }
-        SitecoreInitialiseArgument HydrateCommand(Command command);
-    }
-}
