+++ src/Dimmy.Sitecore.Plugin/Versions/9.3.0/SitecoreInitialiseArgument.cs
@@ -1,0 +1,6 @@
+﻿namespace Dimmy.Sitecore.Plugin.Versions._9._3._0
+{
+    public class SitecoreInitialiseArgument : Sitecore.Plugin.SitecoreInitialiseArgument
+    {
+    }
+}
